Home_Search/backend/requirements.txt
Flask
flask-cors
pymongo
dnspython
python-dotenv
